+++ src/main/java/com/ecommerce/studentmarket/student/cart/exceptions/CartItemProductOutOfStockException.java
@@ -3,7 +3,7 @@ package com.ecommerce.studentmarket.student.cart.exceptions;
 import com.ecommerce.studentmarket.student.cart.domains.CartItemIdDomain;
 
 public class CartItemProductOutOfStockException extends RuntimeException {
-    public CartItemProductOutOfStockException(String tenSP, Integer soLuong) {
+    public CartItemProductOutOfStockException(String tenSP, Long soLuong) {
         super("Sản phẩm "+ tenSP + " không đủ số lượng trong kho để thêm. Chỉ còn lại "+ soLuong + " trong kho.");
     }
 }
